backend/requirements-fastapi.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.36
httpx==0.27.2
PyJWT==2.9.0
numpy==1.26.4
